+++ include/boost/atomic/detail/ops_gcc_ppc_common.hpp
@@ -47,17 +47,17 @@ struct gcc_ppc_operations_base
 #if defined(__powerpc64__) || defined(__PPC64__)
         if (order == memory_order_seq_cst)
             __asm__ __volatile__ ("sync" ::: "memory");
-        else if ((order & memory_order_release) != 0)
+        else if ((static_cast< unsigned int >(order) & static_cast< unsigned int >(memory_order_release)) != 0u)
             __asm__ __volatile__ ("lwsync" ::: "memory");
 #else
-        if ((order & memory_order_release) != 0)
+        if ((static_cast< unsigned int >(order) & static_cast< unsigned int >(memory_order_release)) != 0u)
             __asm__ __volatile__ ("sync" ::: "memory");
 #endif
     }
 
     static BOOST_FORCEINLINE void fence_after(memory_order order) BOOST_NOEXCEPT
     {
-        if ((order & (memory_order_consume | memory_order_acquire)) != 0)
+        if ((static_cast< unsigned int >(order) & (static_cast< unsigned int >(memory_order_consume) | static_cast< unsigned int >(memory_order_acquire))) != 0u)
             __asm__ __volatile__ ("isync" ::: "memory");
     }
 };
